Allow mobs to search through every plant in their plot

+++ src/main/java/com/lovetropics/minigames/common/content/biodiversity_blitz/entity/impl/BbPillagerEntity.java
@@ -95,8 +95,8 @@ public class BbPillagerEntity extends Pillager implements BbMobEntity {
     }
 
     @Override
-    public double aiSpeed() {
-        return 0.65;
+    public float aiSpeed() {
+        return 0.65f;
     }
 
     @Override
